pyproject.toml
[project]
name = "competitor-analysis"
version = "0.1.0"
description = "Social media competitor analysis CLI tool for Kolif Agency"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "openpyxl>=3.1.0",
    "streamlit>=1.56.0",
]

[project.scripts]
competitor-analysis = "competitor_analysis.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/competitor_analysis"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "respx>=0.23.0",
]
